// src/resources/campus/inner.ts
import z from "zod";

export const intraInnerCampusSchema = z.looseObject({
	id: z.number(),
	name: z.string(),
	time_zone: z.string(),
	language: z.looseObject({
		id: z.number(),
		name: z.string(),
		identifier: z.string(),
		created_at: z.coerce.date(),
		updated_at: z.coerce.date(),
	}),
	users_count: z.number(),
	vogsphere_id: z.number().nullable(),
	country: z.string(),
	address: z.string(),
	zip: z.string(),
	city: z.string(),
	website: z.string(),
	facebook: z.string(),
	twitter: z.string(),
	active: z.boolean(),
	public: z.boolean(),
	email_extension: z.string().nullable(),
	default_hidden_phone: z.boolean(),
});